% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editable.R
\name{editable}
\alias{editable}
\title{Create an HTML table widget using the DataTables library}
\usage{
editable(data, editType = NULL, editAttribs = NULL, rownames = F,
  checkboxSelect = F, callback = "", ...)
}
\arguments{
\item{data}{a data frame.}

\item{editType}{A list containing the types of the column editors. The list
must be named by the corresponding column name. Available options are 'text'
for a text input element and 'select' for a select input element.}

\item{editAttribs}{A list containing the options of the column editor elements.
The list must be named by the corresponding column name.
In case of a text input, the option itself must be a list containing the entry
placeholder, which defines the placeholder of the textinput.
In case of a select input, the option itself must be a list containing the vector
named options, which defines the available options of the select input.}

\item{rownames}{Show rownames.}

\item{checkboxSelect}{Select rows using checkboxes.}

\item{callback}{String containing Javascript code to be called during init.}

\item{...}{Other arguments passed to the DataTable() constructor in Javascript.}
}
\description{
This function creates an HTML widget to display a data frame
using the JavaScript library DataTables. The resulting table
may be editable by text inputs and select inputs.
}
\seealso{
editableOutput
}
