% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editable.R
\name{editableOutput}
\alias{editableOutput}
\title{Create an editable HTML table widget output using the DataTables library}
\usage{
editableOutput(outputId, width = "100\%", height = "auto")
}
\arguments{
\item{width}{the width of the table container.}

\item{height}{the height of the table container.}

\item{id}{output variable to read the table from.}
}
\description{
This function creates an HTML widget output to display a data frame 
using the JavaScript library DataTables.
}
\seealso{
editable
}
