% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editable.R
\name{renderEditable}
\alias{renderEditable}
\title{Render an HTML table widget using the DataTables library}
\usage{
renderEditable(expr, env = parent.frame(), quoted = F, ...)
}
\arguments{
\item{expr}{an expression to create a table widget (normally via editable()),
or a data frame to be passed to editable() to create a table widget.}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with quote())?
This is useful if you want to save an expression in a variable.}

\item{...}{ignored when expr returns a table widget, and passed as
additional arguments to editable() when expr returns a data frame}
}
\description{
This function renders an HTML widget to display a data frame
using the JavaScript library DataTables. The resulting table
may be editable by text inputs and select inputs.
}
\seealso{
editable
}
